Design refactorings to extract planet construction, improve readability
