+++ public/index.html
@@ -26,7 +26,7 @@
 
     #pageContainer {
       position: relative;
-      background-color: transparent;
+      background-color: black;
       margin: 10px;
       padding: 17px;
       border: solid white;
