+++ src/startscene.js
@@ -8,16 +8,18 @@ import { StandardMaterial } from "@babylonjs/core/Materials/standardMaterial"
 import { TrailMesh } from "@babylonjs/core/Meshes/trailMesh"
 import { Scalar } from "@babylonjs/core/Maths/math.scalar"
 import { GlowLayer } from "@babylonjs/core/Layers/glowLayer"
-import { Animation } from "@babylonjs/core/Animations/animation"
 import { ArcRotateCamera } from "@babylonjs/core/Cameras/arcRotateCamera";
 import { StarfieldProceduralTexture } from "@babylonjs/procedural-textures/starfield/starfieldProceduralTexture";
 
 // imported for side-effect only
 import "@babylonjs/core/Helpers/sceneHelpers"
 
+// factories
+import AstroFactory from "./astroFactory";
+
+// assets
 import distortTexture from "../assets/textures/distortion.png";
-import rockTextureN from "../assets/textures/rockn.png";
-import rockTexture from "../assets/textures/rock.png";
+
 
 function createStar(scene) {
     let starDiam = 16;
@@ -30,29 +32,6 @@ function createStar(scene) {
     star.rotation = Vector3.Zero();
 
     return star;
-}
-
-function createPlanet(opts, scene) {
-    let planet = MeshBuilder.CreateSphere(opts.name, { diameter: 1 }, scene);
-    let mat = new StandardMaterial(planet.name + "-mat", scene);
-    mat.diffuseColor = mat.specularColor = mat.emissiveColor = opts.color;
-    mat.specularPower = 0;
-    if (opts.rocky === true) {
-        mat.bumpTexture = new Texture(rockTextureN, scene);
-        mat.diffuseTexture = new Texture(rockTexture, scene);
-    }
-    else {
-        mat.diffuseTexture = new Texture(distortTexture, scene);
-    }
-
-    planet.material = mat;
-    planet.scaling.setAll(opts.scale);
-    planet.position.x = opts.posRadius * Math.sin(opts.posRadians);
-    planet.position.z = opts.posRadius * Math.cos(opts.posRadians);
-    planet.orbitOptions = opts;
-    planet.orbitAnimationOvserver = createAndStartOrbitAnimation(planet, scene);
-
-    return planet;
 }
 
 function setupEnvironment(scene) {
@@ -74,7 +53,7 @@ function setupEnvironment(scene) {
 }
 
 function populatePlanetarySystem(scene) {
-    let planets = [];
+    
     let hg = {
         name: "hg",
         posRadians: Scalar.RandomRange(0, 2 * Math.PI),
@@ -115,53 +94,20 @@ function populatePlanetarySystem(scene) {
         color: new Color3(0, 0.3, 1),
         rocky: false
     };
-    planets.push(createPlanet(hg, scene));
-    planets.push(createPlanet(aphro, scene));
-    planets.push(createPlanet(tellus, scene));
-    planets.push(createPlanet(ares, scene));
-    planets.push(createPlanet(zeus, scene));
+    const planetData = [ hg, aphro, tellus, ares, zeus ];
+    const planets = [];
+    planetData.forEach(p => {
+        const planet = AstroFactory.createPlanet(p, scene);
+        AstroFactory.createAndStartOrbitAnimation(planet, scene);
+        planet.computeWorldMatrix(true);
+        let planetTrail = new TrailMesh(planet.name + "-trail", planet, scene, .1, planet.orbitOptions.orbitalCircum, true);
+        let trailMat = new StandardMaterial(planetTrail.name + "-mat", scene);
+        trailMat.emissiveColor = trailMat.specularColor = trailMat.diffuseColor = planet.orbitOptions.color;
+        planetTrail.material = trailMat;
+        planets.push(planet);
+    });    
+
     return planets;
-}
-
-function createSpinAnimation() {
-    let orbitAnim = new Animation("planetspin", "rotation.y", 30, Animation.ANIMATIONTYPE_FLOAT, Animation.ANIMATIONLOOPMODE_CYCLE);
-    const keyFrames = [];
-    keyFrames.push({
-        frame: 0,
-        value: 0
-    });
-
-    keyFrames.push({
-        frame: 60,
-        value: Scalar.TwoPi
-    });
-
-    orbitAnim.setKeys(keyFrames);
-    return orbitAnim;
-}
-
-function createAndStartOrbitAnimation(planet, scene) {
-    const Gm = 6672.59 * 0.07;
-    const opts = planet.orbitOptions;
-    const rCubed = Math.pow(opts.posRadius, 3);
-    const period = Scalar.TwoPi * Math.sqrt(rCubed / Gm);
-    const v = Math.sqrt(Gm / opts.posRadius);
-    const w = v / period;
-    const circum = Scalar.TwoPi * opts.posRadius;
-    let angPos = opts.posRadians;
-
-    planet.computeWorldMatrix(true);
-    let planetTrail = new TrailMesh(planet.name + "-trail", planet, scene, .1, circum, true);
-    let trailMat = new StandardMaterial(planetTrail.name + "-mat", scene);
-    trailMat.emissiveColor = trailMat.specularColor = trailMat.diffuseColor = opts.color;
-    planetTrail.material = trailMat;
-
-    let preRenderObsv = scene.onBeforeRenderObservable.add(sc => {
-        planet.position.x = opts.posRadius * Math.sin(angPos);
-        planet.position.z = opts.posRadius * Math.cos(angPos);
-        angPos = Scalar.Repeat(angPos + w, Scalar.TwoPi);
-    });
-    return preRenderObsv;
 }
 
 export default function createStartScene(engine) {
@@ -175,7 +121,7 @@ export default function createStartScene(engine) {
     const star = that.star = createStar(scene);
     const planets = that.planets = populatePlanetarySystem(scene);
     that.camera = new ArcRotateCamera("camera1", camAlpha, camBeta, camDist, camTarget, scene);
-    const spinAnim = createSpinAnimation();
+    const spinAnim = AstroFactory.createSpinAnimation();
     star.animations.push(spinAnim);
     scene.beginAnimation(star, 0, 60, true);
